Joining and Leaving Town of Salem games: move actions from commands to reaction menus

+++ commands/s/kill.ts
@@ -28,9 +28,9 @@ module.exports = new Command({
                     .setTimestamp();
                     //TO-DO: Add game info (ie players with their respective roles)
                 game.announcements.delete();
-                game.mafia.delete();
-                game.jail.delete();
-                game.graveyard.delete();
+                game.mafia && game.mafia.delete();
+                game.jail && game.jail.delete();
+                game.graveyard && game.graveyard.delete();
                 game.category.delete();
                 game.origin.send(end);
                 game.reset();
